feature(sunjx): Organize the code
